refactor: split pi-ref-tools into maintainable modules
